feat: Integrate the architecture.

+++ API-Blitz/src/main/java/org/example/apiblitz/controller/TestCaseController.java
@@ -1,13 +1,17 @@
 package org.example.apiblitz.controller;
 
+import io.jsonwebtoken.Claims;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.example.apiblitz.model.NextSchedule;
 import org.example.apiblitz.model.ResetTestCase;
 import org.example.apiblitz.model.TestCase;
 import org.example.apiblitz.model.UserResponse;
+import org.example.apiblitz.queue.Publisher;
 import org.example.apiblitz.service.TestCaseService;
+import org.example.apiblitz.util.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -17,6 +21,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+//@Profile("Producer")
 @Controller
 @Slf4j
 @RequestMapping("/api/1.0/testCase")
@@ -25,13 +30,18 @@ public class TestCaseController {
 	@Autowired
 	TestCaseService testCaseService;
 
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	@Autowired
+	Publisher publisher;
+
 	@GetMapping
 	public String testCasePage() {
 		return "testCase";
 	}
 
 	@GetMapping(path = "/get")
-//	public ResponseEntity<?> getTestCase(Integer userId) {
 	public ResponseEntity<?> getTestCase(
 			@RequestHeader("Authorization") String authorization) {
 
@@ -69,15 +79,39 @@ public class TestCaseController {
 			return "redirect:/api/1.0/testCase";
 		}
 
-		String accessToken = extractAccessToken(authorization);
-
 		if (bindingResult.hasErrors()) {
 			throw new BindException(bindingResult);
 		}
 
 		try {
-			Integer testCaseId = testCaseService.save(accessToken, testCase);
-			testCaseService.setTestSchedule(testCaseId, testCase);
+			// User ID
+			String accessToken = extractAccessToken(authorization);
+			Claims claims = jwtUtil.parseToken(accessToken);
+			Integer userId = claims.get("userId", Integer.class);
+//
+//			// Category
+//			String category = "TestCase";
+//
+//			// Type
+//			String type = "create/save";
+//
+//			// ID
+//			Integer id = null;
+//
+//			// Test dateTime
+//			LocalDateTime currentDateTime = LocalDateTime.now().withNano(0);
+//			Timestamp testDateTime = Timestamp.valueOf(currentDateTime);
+//
+//			// Content
+//			Object content = testCase;
+
+//			publisher.publishMessage(userId, category, type, id, testDateTime, content);
+
+			Integer testCaseId = testCaseService.save(userId, testCase);
+
+//			testCaseService.setTestSchedule(userId, Integer.parseInt(message), testCase);
+			testCaseService.setTestSchedule(userId, testCaseId, testCase);
+
 		} catch (Exception e) {
 			log.error(e.getMessage());
 		}
@@ -102,13 +136,33 @@ public class TestCaseController {
 			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(userResponse);
 		}
 
-		String accessToken = extractAccessToken(authorization);
-
 		if (bindingResult.hasErrors()) {
 			throw new BindException(bindingResult);
 		}
 
 		try {
+			String accessToken = extractAccessToken(authorization);
+//			Claims claims = jwtUtil.parseToken(accessToken);
+//			Integer userId = claims.get("userId", Integer.class);
+//
+//			// Category
+//			String category = "TestCase";
+//
+//			// Type
+//			String type = "update";
+//
+//			// ID
+//			Integer id = null;
+//
+//			// Test dateTime
+//			LocalDateTime currentDateTime = LocalDateTime.now().withNano(0);
+//			Timestamp testDateTime = Timestamp.valueOf(currentDateTime);
+//
+//			// Content
+//			Object content = resetTestCase;
+//
+//			publisher.publishMessage(userId, category, type, id, testDateTime, content);
+
 //			testCaseService.update(resetTestCase);
 			testCaseService.update(accessToken, resetTestCase);
 			return ResponseEntity
@@ -121,9 +175,41 @@ public class TestCaseController {
 	}
 
 	@DeleteMapping(path = "/delete")
-	public ResponseEntity<?> deleteTestCase(Integer testCaseId) {
+	public ResponseEntity<?> deleteTestCase(
+			@RequestHeader("Authorization") String authorization,
+			Integer testCaseId) {
+
+		UserResponse userResponse = new UserResponse();
+
+		if (authorization == null || !authorization.startsWith("Bearer ")) {
+			userResponse.setError("Invalid or missing Bearer token");
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(userResponse);
+		}
 
 		try {
+			// User ID
+//			String accessToken = extractAccessToken(authorization);
+//			Claims claims = jwtUtil.parseToken(accessToken);
+//			Integer userId = claims.get("userId", Integer.class);
+//
+//			// Category
+//			String category = "TestCase";
+//
+//			// Type
+//			String type = "delete";
+//
+//			// ID
+//			Integer id = testCaseId;
+//
+//			// Test dateTime
+//			LocalDateTime currentDateTime = LocalDateTime.now().withNano(0);
+//			Timestamp testDateTime = Timestamp.valueOf(currentDateTime);
+//
+//			// Content
+//			Object content = null;
+//
+//			publisher.publishMessage(userId, category, type, id, testDateTime, content);
+
 			testCaseService.delete(testCaseId);
 			return ResponseEntity
 					.ok()
